+++ tests/benchmark/benchmarks.sh
@@ -66,4 +66,14 @@ elif [ "$BM_TYPE" = "bm-spaces" ] ; then
     echo spaces_fp16
     echo spaces_fp64
     echo spaces_bf16
+elif [ "$BM_TYPE" = "bm-spaces-fp32" ] ; then
+    echo spaces_fp32
+elif [ "$BM_TYPE" = "bm-spaces-fp64" ] ; then
+    echo spaces_fp64
+elif [ "$BM_TYPE" = "bm-spaces-bf16" ] ; then
+    echo spaces_bf16
+elif [ "$BM_TYPE" = "bm-spaces-fp16" ] ; then
+    echo spaces_fp16
+elif [ "$BM_TYPE" = "bm-spaces-int8" ] ; then
+    echo spaces_int8
 fi
