+++ tests/unit/test_allocator.cpp
@@ -10,7 +10,7 @@
 #include "VecSim/memory/vecsim_base.h"
 #include "VecSim/algorithms/brute_force/brute_force_single.h"
 #include "VecSim/algorithms/hnsw/hnsw_single.h"
-#include "test_utils.h"
+#include "unit_test_utils.h"
 #include "VecSim/utils/serializer.h"
 #include "VecSim/index_factories/hnsw_factory.h"
 
@@ -83,7 +83,7 @@ TEST_F(AllocatorTest, test_nested_object) {
 template <typename index_type_t>
 class IndexAllocatorTest : public ::testing::Test {};
 
-// DataTypeSet, TEST_DATA_T and TEST_DIST_T are defined in test_utils.h
+// DataTypeSet, TEST_DATA_T and TEST_DIST_T are defined in unit_test_utils.h
 
 TYPED_TEST_SUITE(IndexAllocatorTest, DataTypeSet);
