+++ tests/unit/test_hnsw_tiered.cpp
@@ -6,7 +6,7 @@
 #include <string>
 #include <array>
 
-#include "test_utils.h"
+#include "unit_test_utils.h"
 #include "mock_thread_pool.h"
 
 #include <thread>
